Moving the rate measurer to its own file to be shared between scenarios
